pyproject.toml
[tool.poetry]
name = "udev-monitor"
version = "0.1.0"
description = ""
authors = ["inahym196 <inahym196@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pyudev = "^0.22.0"
pyserial = "^3.5"
pyserial-asyncio = "^0.5"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
